requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Agent Framework & LLM
langgraph
langchain
langchain-openai
openai

# MCP Protocol
mcp

# Database
neo4j==5.17.0

# Configuration & Validation
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Async & HTTP
httpx==0.26.0
aiofiles==23.2.1

# Repository Management
GitPython==3.1.41

# Utilities
python-multipart==0.0.6
typing-extensions

# Development & Testing
pytest==7.0.0
pytest-asyncio==0.23.4
black==24.1.1
ruff==0.2.1
